- Fixed audio entry panel crash when opening a midi in linux - Fixed doom sound -> wav conversion with gcc - Added MIDI playback via fluidsynth in linux (doesn't seem to want to work yet, though)
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@687 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/Conversions.cpp
@@ -6,17 +6,17 @@
 // Some structs for wav conversion
 struct wav_chunk_t {
 	char id[4];
-	long size;
+	uint32_t size;
 };
 
 struct wav_fmtchunk_t {
 	wav_chunk_t header;
-	short tag;
-	short channels;
-	long samplerate;
-	long datarate;
-	short blocksize;
-	short bps;
+	uint16_t tag;
+	uint16_t channels;
+	uint32_t samplerate;
+	uint32_t datarate;
+	uint16_t blocksize;
+	uint16_t bps;
 };
 
 struct dsnd_header_t {
